refactor(www): group codeOptions UI into a create/code-options module
The codeOptions UI is the one customizer axis that's a modal + live
preview rather than a single inline panel, so it earns a folder. Split it
into focused files and extract the live preview into its own unit:

  create/code-options/
    index.ts      → CodeOptionsDialog
    dialog.tsx    → footer button + modal shell
    controls.tsx  → the two controls (was code-config.tsx)
    preview.tsx   → live /r/button preview (was inline in the dialog)

The engine (publisher/code-options.ts) stays put — it's shared with the
route bundle and must remain import-free.

Also regenerates showcase-bundle.ts: the bundled copy of code-options.ts
was a line stale (a regex wrapped by the formatter after the previous
build), which the registry-drift CI job would have flagged.
