+++ www/src/modules/create/code-options/preview.tsx
@@ -1,62 +1,24 @@
 /**
- * Footer entry point for `codeOptions`: a "Code style" button that opens a
- * split modal — the code-style controls on the left, a live preview of the
- * exported button component on the right. The preview fetches the real
- * `/r/button` registry endpoint (same publisher + oxfmt the user gets on
- * install), debounced, so every toggle shows the actual resulting source.
+ * Live preview of the exported button component. Fetches the real `/r/button`
+ * registry endpoint (same publisher + oxfmt the user gets on install),
+ * debounced and same-origin, so every codeOptions toggle shows the actual
+ * resulting source. Highlighting reuses the docs `DynamicPre` (synchronous
+ * shiki).
  */
 
 import { useEffect, useMemo, useState } from 'react'
-import { CheckIcon, Code2Icon, CopyIcon } from 'lucide-react'
+import { CheckIcon, CopyIcon } from 'lucide-react'
 
 import { useCopyToClipboard } from '@/hooks/use-copy-to-clipboard'
 import { Button } from '@/registry/ui/button'
-import { Dialog, DialogContent } from '@/registry/ui/dialog'
-import { Overlay } from '@/registry/ui/overlay'
 import { DynamicPre } from '@/modules/docs/dynamic-pre'
 
-import { CodeConfig } from './code-config'
-import { useDesignSystem } from './preset'
-import { encodePreset } from './preset/codec'
+import { useDesignSystem } from '../preset'
+import { encodePreset } from '../preset/codec'
 
 const PREVIEW_FILE = 'components/ui/button.tsx'
 
-export function CodeOptionsDialog() {
-  return (
-    <Dialog>
-      <Button variant="default" size="sm" className="w-full">
-        <Code2Icon data-icon-start="" />
-        Code style
-      </Button>
-      <Overlay
-        type="modal"
-        mobileType={null}
-        modalProps={{
-          className: 'h-[min(38rem,82vh)] w-[min(64rem,92vw)] sm:max-w-none',
-        }}
-      >
-        <DialogContent
-          showCloseButton
-          aria-label="Code style"
-          className="gap-0 overflow-hidden p-0"
-        >
-          <div className="flex h-full min-h-0 flex-col md:flex-row">
-            <div className="flex w-full shrink-0 flex-col overflow-y-auto border-b p-4 md:w-80 md:border-r md:border-b-0">
-              <h2 className="text-sm font-medium">Code style</h2>
-              <p className="mt-0.5 mb-4 text-xs text-fg-muted">
-                Shape the exported code to match your codebase.
-              </p>
-              <CodeConfig />
-            </div>
-            <CodePreview />
-          </div>
-        </DialogContent>
-      </Overlay>
-    </Dialog>
-  )
-}
-
-function CodePreview() {
+export function CodeOptionsPreview() {
   const { designSystem } = useDesignSystem()
   const encoded = useMemo(() => encodePreset(designSystem), [designSystem])
